fix, refactor: 비동기 관련 오류 수정
프로젝트 내 비동기 관련 함수 오류 수정

-   ** utils/cg_ui_handlers.py 파일 내 reasoning_handler 호출 관련 오류 수정 **
-   ** 기존 분리된 비동기 함수 메서드 이동 ( 리팩터링 ) **
