+++ utils/cg_auto_handlers.py
@@ -6,145 +6,6 @@ from prompt_toolkit.formatted_text import HTML
 from utils.prompt_select import prompt_select
 from typing import Optional, Dict
 
-async def _plan_step(self, task_node: Dict) -> Optional[Dict]:
-    """Helper for the 'Plan' phase. Decides which tool to use based on task and memory."""
-    self.console.print(f"\n[bold cyan]Phase: Planning[/bold cyan]")
-    self.console.print(f"  - Task: [yellow]{task_node['description']}[/yellow]")
-    
-    # 1. Fetch relevant memories
-    with self.console.status("[green]Searching knowledge base for planning context..."):
-        query_text = task_node['description']
-        query_results = await asyncio.to_thread(
-            self.memory_manager.chroma_manager.query_by_text,
-            query_text, n_results=3
-        )
-    
-    memory_context_parts = []
-    if query_results and query_results['ids'] and query_results['ids'][0]:
-        for i, doc in enumerate(query_results['documents'][0]):
-            memory_context_parts.append(f"- {doc}")
-    memory_context = "\n".join(memory_context_parts) if memory_context_parts else "No relevant memories found."
-    self.console.print(f"  - Found {len(memory_context_parts)} relevant memories for context.")
-
-    # 2. Fetch available tools
-    with self.console.status("[green]Fetching available tools from MCP servers..."):
-        all_tools_by_server = {}
-        server_names = self.mcp_manager.get_available_servers()
-        for server_name in server_names:
-            client = self.mcp_manager.create_client(server_name)
-            if not client: continue
-            try:
-                async with client:
-                    if client.is_connected:
-                        tools = await client.list_tools()
-                        all_tools_by_server[server_name] = [
-                            {"name": tool.name, "description": tool.description, "inputSchema": tool.inputSchema} for tool in tools
-                        ]
-            except Exception as e:
-                self.console.print(f"[yellow]Warning: Could not fetch tools from '{server_name}': {e}[/yellow]")
-    
-    if not all_tools_by_server:
-        self.console.print("[bold red]No tools available from any MCP server. Cannot plan.[/bold red]")
-        return None
-
-    # 3. LLM call to decide action
-    all_tools_json_str = json.dumps(all_tools_by_server, indent=2)
-    prompt_for_full_action = f"""
-    You are an expert CTF assistant responsible for selecting the best tool for a given task.
-
-    Current Task:
-    ---
-    {task_node['description']}
-    ---
-
-    Relevant Information from Memory:
-    ---
-    {memory_context}
-    ---
-
-    Available Tool Servers and their Tools:
-    ```json
-    {all_tools_json_str}
-    ```
-
-    Based on the task and your memory, select the single most appropriate server and tool to execute.
-    Your output MUST be a single JSON object with "server_name", "tool_name", and "arguments" keys.
-    If no tool is suitable, return a JSON object with "server_name" set to "none".
-    """
-    
-    with self.console.status("[green]LLM is planning the next action..."):
-        action_json_str = await asyncio.to_thread(
-            self.reasoningAgent.send_message,
-            prompt_for_full_action, self.test_reasoning_session_id, json_mode=True
-        )
-
-    try:
-        # LLM 응답에 포함될 수 있는 Markdown 코드 블록 제거
-        cleaned_json_str = re.sub(r'```json\s*|\s*```', '', action_json_str).strip()
-        action = json.loads(cleaned_json_str)
-        self.console.print(f"  - Planned Action: Run [cyan]'{action.get('tool_name')}'[/cyan] on server [cyan]'{action.get('server_name')}'[/cyan]")
-        return action
-    except json.JSONDecodeError:
-        # JSON 파싱 실패 시, 정규식을 최후의 수단으로 사용
-        json_match = re.search(r'\{.*\}', action_json_str, re.DOTALL)
-        if json_match:
-            try:
-                action = json.loads(json_match.group(0))
-                self.console.print(f"  - Planned Action (extracted from response): Run [cyan]'{action.get('tool_name')}'[/cyan] on server [cyan]'{action.get('server_name')}'[/cyan]")
-                return action
-            except json.JSONDecodeError as e:
-                pass # Fall through to error
-        
-        self.console.print(f"[bold red]Failed to parse action from LLM: Invalid JSON[/bold red]")
-        self.console.print(f"LLM Response: {action_json_str}", style="red")
-        return None
-
-async def _execute_step(self, action: dict) -> tuple[Optional[str], bool]:
-    """
-    Helper for the 'Execute' phase. Runs the planned tool command.
-    Returns a tuple of (output, is_error).
-    """
-    self.console.print(f"\n[bold cyan]Phase: Executing[/bold cyan]")
-    server_name = action.get("server_name")
-    tool_name = action.get("tool_name")
-    arguments = action.get("arguments")
-
-    self.console.print(f"  - Executing tool [yellow]'{tool_name}'[/yellow] on server [yellow]'{server_name}'[/yellow]")
-    self.console.print(f"  - Arguments: [dim]{json.dumps(arguments)}[/dim]")
-    
-    client = self.mcp_manager.create_client(server_name)
-    if not client:
-        error_msg = f"Failed to create client for server '{server_name}'."
-        self.console.print(f"[bold red]{error_msg}[/bold red]")
-        return error_msg, True
-    
-    tool_output_text = None
-    try:
-        async with client:
-            if not client.is_connected:
-                error_msg = f"MCP Client could not connect to '{server_name}'."
-                self.console.print(f"[bold red]{error_msg}[/bold red]")
-                return error_msg, True
-            
-            with self.console.status(f"[green]Running '{tool_name}'..."):
-                tool_results_list = await client.call_tool(tool_name, arguments)
-                if not tool_results_list:
-                    tool_output_text = f"Tool '{tool_name}' executed but returned no result."
-                    self.console.print(f"[yellow]{tool_output_text}[/yellow]")
-                else:
-                    tool_output_text = "\n".join([result.text for result in tool_results_list])
-                
-                self.console.print("  - Tool execution complete.")
-                self.console.print("[bold green]--- Tool Output ---[/bold green]")
-                self.console.print(f"[dim]{(tool_output_text or '')[:1000]}...[/dim]") # Print snippet
-                self.console.print("[bold green]-------------------[/bold green]")
-                return tool_output_text, False
-
-    except Exception as e:
-        tool_output_text = f"Error during tool execution: {e}"
-        self.console.print(f"[bold red]{tool_output_text}[/bold red]")
-        return tool_output_text, True # Return error message and error flag
-
 async def run_single_auto_step(self, interactive: bool):
     """
     This function now orchestrates the memory-centric Plan-Execute-Memorize-Reason loop and correctly handles async calls..
@@ -160,18 +21,18 @@ async def run_single_auto_step(self, interactive: bool):
         return
 
     # 2. PLAN
-    action = await _plan_step(self, current_task_node)
+    action = await self._plan_step(current_task_node)
 
     if not action or action.get("server_name") in ['none', None]:
         self.console.print("[yellow]LLM decided not to execute a tool. Proceeding to reason on existing knowledge.[/yellow]")
         # Even if no tool is run, we reason about the task with existing knowledge.
-        await self.reasoning_handler(current_task_node) 
+        await self.reasoning_handler(current_task_node)
         self.console.print("\n[bold green]CTT after reasoning step:[/bold green]")
         self.console.print(self.ctt_manager.get_display_string())
         return
 
     # 3. EXECUTE
-    tool_output, is_error = await _execute_step(self, action)
+    tool_output, is_error = await self._execute_step(action)
 
     # 4. MEMORIZE
     self.console.print(f"\n[bold cyan]Phase: Memorizing[/bold cyan]")
