+++ utils/cg_ui_handlers.py
@@ -152,7 +152,7 @@ def input_handler(self) -> str:
                 user_input, source=source
             )
             ## (2) pass the summarized information to the reasoning session.
-            reasoning_response = self.reasoning_handler(parsed_input)
+            reasoning_response = asyncio.run(self.reasoning_handler(parsed_input))
             # self.step_reasoning = reasoning_response
 
         ## (3) print the results
@@ -215,7 +215,7 @@ def input_handler(self) -> str:
         self.log_conversation("user", "todo")
         ## (1) ask the reasoning session to analyze the current situation, and list the top sub-tasks
         with self.console.status("[bold green] CTFGPT Thinking...") as status:
-            reasoning_response = self.reasoning_handler(self.prompts.ask_todo)
+            reasoning_response = asyncio.run(self.reasoning_handler(self.prompts.ask_todo))
             self.step_reasoning = reasoning_response 
             ## (2) pass the sub-tasks to the test_generation session.
             message = self.prompts.todo_to_command + "\n" + reasoning_response
@@ -250,7 +250,7 @@ def input_handler(self) -> str:
         self.log_conversation("user", user_input)
         ## (2) pass the information to the reasoning session.
         with self.console.status("[bold green] CTFGPT Thinking...") as status:
-            response = self.reasoning_handler(self.prompts.discussion + user_input)
+            response = asyncio.run(self.reasoning_handler(self.prompts.discussion + user_input))
         ## (3) print the results
         self.console.print("CTFGPT:\n", style="bold green")
         self.console.print(response + "\n", style="yellow")
